Prevent strtold from wrongfully capturing operations
If you passed in an equation with no spaces (e.g. "2+2") strtold
would capture the "+" as an indication of the sign of the second
number and so hence only two NUM tokens were lexed.
The new logic here assumes that if a + or - preceeds and follows a
number, then it is probably an operation.